change to C++ way: IndexWhereLowerThan to index_if, intersect1d

+++ matching_lines.hpp
@@ -16,7 +16,7 @@ namespace cv { namespace separableFundamentalMatrix
 
     struct line_info
     {
-        vector<int> matching_indexes;
+        vector<int64> matching_indexes;
         Point3f line_eq_abc;
         Point3f line_eq_abc_norm;
         Point2f bottom_left_edge_point;
